+++ test/commands/support/consistency/strongly_consistent_event_handler.ex
@@ -1,29 +1,26 @@
 defmodule Commanded.Commands.StronglyConsistentEventHandler do
   use Commanded.Event.Handler,
-    application: Commanded.DefaultApp,
+    application: Commanded.Commands.ConsistencyApp,
     name: "StronglyConsistentEventHandler",
     consistency: :strong
 
-  alias Commanded.Commands.{
-    ConsistencyAggregateRoot,
-    ConsistencyRouter
-  }
+  alias Commanded.Commands.{ConsistencyAggregateRoot, ConsistencyApp}
+  alias ConsistencyAggregateRoot.{ConsistencyCommand, ConsistencyEvent, DispatchRequestedEvent}
 
-  alias ConsistencyAggregateRoot.{
-    ConsistencyCommand,
-    ConsistencyEvent,
-    DispatchRequestedEvent
-  }
-
-  # handle event by dispatching a command
+  @doc """
+  Dispatch a command with consistency `:strong` after an optional delay.
+  """
   def handle(%DispatchRequestedEvent{uuid: uuid, delay: delay}, _metadata) do
     :timer.sleep(delay)
 
     command = %ConsistencyCommand{uuid: uuid, delay: delay}
 
-    ConsistencyRouter.dispatch(command, application: Commanded.DefaultApp, consistency: :strong)
+    ConsistencyApp.dispatch(command, consistency: :strong)
   end
 
+  @doc """
+  Block for a requested delay.
+  """
   def handle(%ConsistencyEvent{delay: delay}, _metadata) do
     :timer.sleep(delay)